src/pyproject.toml
[tool.poetry]
name = "fantasy-ai"
version = "0.1.0"
description = ""
authors = ["James Spillmann <james@linda-rick.com>"]
readme = "README.md"

[tool.poetry.scripts]
run-rankings = "fantasy_ai.rankings_service.main:main"

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.10.1"
bs4 = "^0.0.2"
asyncpg = "^0.29.0"
gnews = "^0.3.7"
langchain-openai = "^0.1.20"
poetry-dotenv-plugin = "^0.2.0"
langchain-chroma = "^0.1.2"
langchain = "^0.2.12"
langchainhub = "^0.1.20"
google-cloud-firestore = "^2.17.0"
pytz = "^2024.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
path = ".venv"

